Fix t/03database.t tests
The original code tested wheter the DB connection could be setup and not
if the code actually did what is what supposed to do. Refactor it so it
tests code written in Act::Database.

This change also eliminates the need for Act::Config as the functions
don't need any configuration, just a database handle which
Test::MockObject now provides to us.